requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
motor==2.5.1
pandas==1.3.3
python-dotenv==0.19.0
openai==0.27.0 